requirements.txt
Django
django-ckeditor
django-import-export
django-admin-thumbnails